PythonAssignments/assignement9_sqlAlchemy/assignment1_sqlAlchemy/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
pydantic
